Move DependencyInjection to Application namespace
