+++ tests/Application/Bootstrap/BootstrapTest.php
@@ -15,8 +15,8 @@ namespace Desperado\ServiceBus\Tests\Application\Bootstrap;
 
 use Desperado\ServiceBus\Application\Bootstrap;
 use function Desperado\ServiceBus\Common\removeDirectory;
-use Desperado\ServiceBus\DependencyInjection\Compiler\TaggedMessageHandlersCompilerPass;
-use Desperado\ServiceBus\DependencyInjection\Extensions\ServiceBusExtension;
+use Desperado\ServiceBus\Application\DependencyInjection\Compiler\TaggedMessageHandlersCompilerPass;
+use Desperado\ServiceBus\Application\DependencyInjection\Extensions\ServiceBusExtension;
 use Desperado\ServiceBus\Infrastructure\Storage\SQL\DoctrineDBAL\DoctrineDBALAdapter;
 use PHPUnit\Framework\TestCase;
